refactor: move the rest of the components to the Wintry folder
