+++ src/components/Wintry/Settings/pages/Addon/AddonCollectionManager.tsx
@@ -1,7 +1,7 @@
 import fuzzysort from "fuzzysort";
-import { useSearchQuery } from "../../../../Search";
+import { useSearchQuery } from "../../../../Wintry/Search";
 import { create, type StoreApi, type UseBoundStore } from "zustand";
-import { type SortOption, type FilterOption, FilterAndSortBar } from "../../../../FilterAndSortBar";
+import { type SortOption, type FilterOption, FilterAndSortBar } from "../../../FilterAndSortBar";
 import type { Addon } from ".";
 
 type FilterAndSortStore<SK extends string, FK extends string> = {
