+++ src/stores/useUpdaterStore.tsx
@@ -2,7 +2,7 @@ import { Card, Text } from "@components/Discord";
 import UpdaterModule, { type UpdateInfo } from "@loader/modules/UpdaterModule";
 import { create } from "zustand";
 import { showAlert } from "@api/alerts";
-import ErrorCard from "@components/ErrorCard";
+import ErrorCard from "@components/Wintry/ErrorCard";
 import { showToast } from "@api/toasts";
 import { Mutex, delay, noop, pick } from "es-toolkit";
 import { t } from "@i18n";
